+++ src/main/java/com/facilio/bmsconsole/commands/AddOrUpdatePurchaseContractCommand.java
@@ -1,9 +1,20 @@
 package com.facilio.bmsconsole.commands;
 
 import com.facilio.beans.ModuleBean;
+import com.facilio.bmsconsole.context.ContractsContext;
+import com.facilio.bmsconsole.context.ContractsContext.ContractType;
+import com.facilio.bmsconsole.context.ContractsContext.Status;
 import com.facilio.bmsconsole.context.ContractsContext.ContractType;
 import com.facilio.bmsconsole.context.PurchaseContractContext;
 import com.facilio.bmsconsole.context.PurchaseContractLineItemContext;
+import com.facilio.bmsconsole.criteria.CriteriaAPI;
+import com.facilio.bmsconsole.criteria.NumberOperators;
+import com.facilio.bmsconsole.modules.DeleteRecordBuilder;
+import com.facilio.bmsconsole.modules.FacilioField;
+import com.facilio.bmsconsole.modules.FacilioModule;
+import com.facilio.bmsconsole.modules.InsertRecordBuilder;
+import com.facilio.bmsconsole.modules.ModuleBaseWithCustomFields;
+import com.facilio.bmsconsole.modules.UpdateRecordBuilder;
 import com.facilio.constants.FacilioConstants;
 import com.facilio.db.criteria.CriteriaAPI;
 import com.facilio.db.criteria.operators.NumberOperators;
@@ -23,6 +34,8 @@ public class AddOrUpdatePurchaseContractCommand implements Command {
 	public boolean execute(Context context) throws Exception {
 		String moduleName = (String) context.get(FacilioConstants.ContextNames.MODULE_NAME);
 		PurchaseContractContext purchaseContractContext = (PurchaseContractContext) context.get(FacilioConstants.ContextNames.RECORD);
+		boolean isContractRevised = (boolean) context.get(FacilioConstants.ContextNames.IS_CONTRACT_REVISED);
+
 		if (purchaseContractContext != null) {
 			ModuleBean modBean = (ModuleBean) BeanFactory.lookup("ModuleBean");
 			FacilioModule module = modBean.getModule(moduleName);
@@ -37,21 +50,41 @@ public class AddOrUpdatePurchaseContractCommand implements Command {
 				throw new Exception("Vendor cannot be empty");
 			}
 			purchaseContractContext.setContractType(ContractType.PURCHASE);
-			if (purchaseContractContext.getId() > 0) {
+			if (!isContractRevised && purchaseContractContext.getId() > 0) {
 				updateRecord(purchaseContractContext, module, fields);
 				
 				DeleteRecordBuilder<PurchaseContractLineItemContext> deleteBuilder = new DeleteRecordBuilder<PurchaseContractLineItemContext>()
 						.module(lineModule)
 						.andCondition(CriteriaAPI.getCondition("PURCHASE_CONTRACT", "purchaseContractId", String.valueOf(purchaseContractContext.getId()), NumberOperators.EQUALS));
 				deleteBuilder.delete();
-			} else {
-				
-				purchaseContractContext.setStatus(PurchaseContractContext.Status.WAITING_FOR_APPROVAL);
-				addRecord(true,Collections.singletonList(purchaseContractContext), module, fields);
+				updateLineItems(purchaseContractContext);
+				addRecord(false,purchaseContractContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+
 			}
-			
-			updateLineItems(purchaseContractContext);
-			addRecord(false,purchaseContractContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+			else if (isContractRevised && purchaseContractContext.getId() > 0) {
+				if(purchaseContractContext.getStatusEnum() == Status.APPROVED) {
+					PurchaseContractContext revisedContract =  purchaseContractContext.clone();
+					addRecord(true,Collections.singletonList(revisedContract), module, fields);
+					purchaseContractContext.setStatus(Status.REVISED);
+					updateRecord(purchaseContractContext, module, fields);
+					updateLineItems(revisedContract);
+					addRecord(false,revisedContract.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+				}
+				else {
+					throw new IllegalArgumentException("Only Approved contracts can be revised");
+				}
+
+			}
+			else {
+				purchaseContractContext.setStatus(PurchaseContractContext.Status.WAITING_FOR_APPROVAL);
+				purchaseContractContext.setRevisionNumber(0);
+				addRecord(true,Collections.singletonList(purchaseContractContext), module, fields);
+				purchaseContractContext.setParentId(purchaseContractContext.getLocalId());
+				updateRecord(purchaseContractContext, module, fields);
+				updateLineItems(purchaseContractContext);
+				addRecord(false,purchaseContractContext.getLineItems(), lineModule, modBean.getAllFields(lineModule.getName()));
+			}
+
 			
 			context.put(FacilioConstants.ContextNames.RECORD, purchaseContractContext);
 		}
